ingestion/producers/database/requirements.txt
pandas
psycopg2-binary
sqlalchemy
confluent-kafka